--- payment-service/src/main/java/shop/genieus/payment/application/dto/RegisterPaymentCommand.java
@@ -1,7 +1,0 @@
-package shop.genieus.payment.application.dto;
-
-public record RegisterPaymentCommand(
-        String paymentKey, // 일단 검증용
-        Long orderId,
-        Long amount // 일단 검증용
-) {}
