+++ payment-service/src/main/java/shop/genieus/payment/application/in/dto/RegisterPaymentCommand.java
@@ -1,0 +1,7 @@
+package shop.genieus.payment.application.in.dto;
+
+public record RegisterPaymentCommand(
+        String paymentKey,
+        Long orderId,
+        Long amount
+) {}
