+++ payment-service/src/main/java/shop/genieus/payment/presentation/rest/dto/request/ProcessPaymentRequest.java
@@ -1,7 +1,7 @@
 package shop.genieus.payment.presentation.rest.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import shop.genieus.payment.application.dto.ProcessPaymentCommand;
+import shop.genieus.payment.application.in.dto.ProcessPaymentCommand;
 
 public record ProcessPaymentRequest(
         @JsonProperty("orderId") Long orderId,
